+++ backend/db/seeds.rb
@@ -16,6 +16,8 @@ User.destroy_all
 User.create!(
   name: "Alice Teacher",
   email: "alice@nku.edu",
+  password: "password",
+  password_confirmation: "password",
   role: :teacher
 )
 
@@ -23,6 +25,8 @@ User.create!(
 User.create!(
   name: "Bob Student",
   email: "bob@nku.edu",
+  password: "password",
+  password_confirmation: "password",
   role: :student
 )
